+++ DMX.Core.Api/Brokers/LabApis/IReverbApiBroker.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ---------------------------------------------------------------
 
-namespace DMX.Core.Api.Brokers.ReverbApis
+namespace DMX.Core.Api.Brokers.LabApis
 {
     public partial interface IReverbApiBroker
     {
